requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.2
gunicorn==20.1.0